Refactor #761 Phase 2 (限定スコープ) + #769: 共有 helper 化と重複 modlog test 削除
## Phase 2 (限定スコープ)

#761 Phase 1 で確立した setupXxxHandler パターンを残り admin test file に
横展開する。Phase 2 候補 (accounts/moderation/captcha/forward_abuse_report/
email、計 26 occurrence) の大半は newTestHandler が返す tuple repo
(userRepo / metaRepo) を使っており Phase 1 と pattern が異なるため、helper
化の旨味があるのは 3 test のみ。

そこで Phase 2 は限定スコープで進める:
- setupDriveFileHandler を drive_emoji_test.go から handler_test.go に
  移動して共有化 (cross-file 再利用可能に)
- setupAbuseReportHandler を新規追加 (handler_test.go に共有 helper として)
- 移行対象:
  - moderation_test.go: TestDeleteAllFilesOfUser_DeletesOnlyTargetUserFiles
    (DriveFileRepo)
  - moderation_test.go: TestUpdateAbuseUserReport_WritesModerationLog
    (AbuseRepo)
  - forward_abuse_report_test.go: TestForwardAbuseUserReport_WritesModerationLog
    (AbuseRepo)

残る 23 occurrence (smoke / tuple-based / ad-hoc stub 系) は意図的に保持。

## #769: 重複 modlog test 削除

abuse_report_notification_test.go の moderation log assertion section に
完全に同等な test 3 ペアが #665 由来で混入していた:

- TestRecipientCreate_WritesModerationLog ≒
  TestAbuseReportNotificationRecipientCreate_WritesModerationLog
- 同 Update / Delete pair

後者 3 件 (subset) を削除し、重複 section header
"// --- moderation log assertions (#665) ---" も解消。前者は assertion が
より強い (Update test に before/after 検証あり) ので残す。

Phase 1a (#768) で helper 化した結果、重複が一目で分かるようになったため
本 PR で同時に消化。

## Closes

- #769

## Phase 2 で意図的に skip した tests

| 種類 | 件数 | 理由 |
|---|---:|---|
| nil-repo smoke | 9 | 1 行 test、helper 化の旨味なし |
| tuple-based userRepo / metaRepo | 11 | newTestHandler が wire 済み、Set 不要 |
| ad-hoc stub (forwarder/fetcher/enqueuer) | 6 | per-test に異なる stub state |

合計 26 occurrence のうち 3 件のみ helper 化。
